Toolbar / Banner Refactoring | Part 1
- Adding a global "Toolbar" / "Banner Window" at the top of the main GUI (right below the "file, edit, view, ... " MenuBar)

- This toolbar will be a persistent / CORE aspect of the main user interface.

- The purpose of this toolbar is to provide a global menu from which the user can access buttons to manage the state of the application.

- This includes things like...
    - Open/Close the sidebar window
    - Open/Close the control window (at the bottom of the app)
    - [POTENTIALLY] Access tools that are related to specific applications---such as the "DEFAULT" / "DRAW" / "BUILD" state of the FDTD Editor Application.

+++ src/app/graph_app/graph_app.cpp
@@ -164,7 +164,6 @@ void GraphApp::display_plots(void)
 
 
     ImGui::EndTabBar();
-        
     } // END "BeginTabBar".
 
     return;
@@ -484,8 +483,8 @@ static constexpr float      WIDGET_WIDTH        = 250.0f;
 void GraphApp::EditorUtility(void)
 {
     // ---------------------------------------------------------------------
-    const int res_x = m_steps.NX.value;   // external grid width
-    const int res_y = m_steps.NY.value;   // external grid height
+    const int res_x = static_cast<int>( m_steps.NX.value );   // external grid width
+    const int res_y = static_cast<int>( m_steps.NY.value );   // external grid height
     sketch.resize(res_x, res_y);
 
     ImPlotHeatmapFlags hm_flags   = 0;
